moved accept header. updated other headers to work the same way
