--- src/Symfony/Component/HttpFoundation/Header/ParsedHeaderInterface.php
@@ -1,15 +1,0 @@
-<?php
-/**
- * Created by PhpStorm.
- * User: yosefderay
- * Date: 6/28/15
- * Time: 1:12 PM
- */
-
-namespace Symfony\Component\HttpFoundation\Header;
-
-
-interface ParsedHeaderInterface
-{
-    public function __toString();
-}
